FIX: PatternFinder getting shape from image
If the image is set before the `find` method-call by the `set_image`
method and when no ROI is given, we have to get the height/width from
the image_gpu (because we no longer have access to the numpy array
called "image"). Note that h/w is switched for image_gpu because OpenCL
stores it that way.